requirements.txt
alembic==1.16.5
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
backports.asyncio.runner==1.2.0
bcrypt==4.3.0
blinker==1.9.0
Brotli==1.1.0
cachetools==6.2.0
catboost==1.2.8
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
contourpy==1.3.2
coverage==7.11.0
cryptography==46.0.3
cssselect2==0.8.0
cycler==0.12.1
exceptiongroup==1.3.0
fastapi==0.116.1
fonttools==4.59.2
gitdb==4.0.12
GitPython==3.1.45
graphviz==0.21
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
python-jose[cryptography]==3.5.0
python-multipart==0.0.9
kiwisolver==1.4.9
lightgbm==4.6.0
llvmlite==0.44.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.6
narwhals==2.3.0
numba==0.61.2
numpy==2.2.6
nvidia-nccl-cu12==2.28.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
plotly==6.3.0
pluggy==1.6.0
protobuf==6.32.0
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pydyf==0.11.0
Pygments==2.19.2
pyparsing==3.2.3
pyphen==0.17.2
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
scikit-learn==1.7.1
scipy==1.15.3
shap==0.48.0
six==1.17.0
slicer==0.0.8
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
streamlit==1.49.1
tenacity==9.1.2
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
tomli==2.2.1
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchdog==6.0.0
weasyprint==66.0
webencodings==0.5.1
xgboost==3.0.5
zopfli==0.2.3.post1
